+++ src/BadgesCommentCell.hpp
@@ -12,12 +12,14 @@ class $modify(BadgesCommentCell, CommentCell) {
         CCNode* m_self;
         CCNode* m_usernameNode;
         CCNode* m_usernameMenu;
+        CCLabelBMFont* m_commentNode;
         CCNode* m_badgeNode;
         CCMenuItemSpriteExtra* m_badgesButton;
         std::vector<alpha::badgify::BadgeInfo*> m_badges;
         std::vector<std::shared_ptr<alpha::badgify::Badge>> m_owningBadges;
         std::unordered_map<alpha::badgify::BadgeInfo*, Ref<CCNode>> m_badgeNodes;
         bool m_canAddBadge = false;
+        cocos2d::ccColor3B m_originalColor = {255, 255, 255};
 
         ~Fields() {
             for (const auto& owningBadge : m_owningBadges) {
@@ -31,6 +33,7 @@ class $modify(BadgesCommentCell, CommentCell) {
     void removeBadge(alpha::badgify::BadgeInfo* info);
     void setupBadges();
     void updateBadges();
+    void updateColor();
     void unregisterTouchesRecursive(CCNode* node);
     void addToBadgeContainer(alpha::badgify::BadgeInfo* info, CCNode* badgeNode);
     void onBadges(CCObject* sender);